requirements.txt
asyncio
discord
python-decouple
mysql-connector-python
twscrape